daisy: add qcamerasvr restart workaround
The cameraservice seems to be broken in some mysterious way and since we do not have sources for the blobs, we must figure out a workaround. The main issue is that when HAL3 is enabled, we have GCAM working but other apps still use HAL3 even if listed in hal1.packagelist/hal3.packagelist. They don't seem to care, so this will enable/disable hal1/hal3 and hopefully provide a workaround. Thanks again to the original commit.

------ Original commit message ------
Unfortunately, s2's mm-qcamera-daemon has an issue where sometimes
it can't receive any valid preview frame at times which causes the
camera to black screen and fail to work until a reboot is
performed. This issue could be due to us using camera HAL3 when
LeEco only inteded for HAL1 to be used as HAL3 is disabled on our
stock ROM.

Without the source code to mm-qcamera-daemon and the camera blobs
we cannot currently fix this issue. A workaround we can do though
to prevent a reboot is to restart the qcamerasvr manually by
adding a setting for it.

Another workaround is to use camera HAL1 as it works perfectly,
especially in certain apps like WhatsApp where switching between
front/rear camera repeatedly causes the blackscreen issue in
camera HAL3.

This commit is co-authored by @zhaoguomanong as he implemented a
better camera reset implementation, added a camera HAL3/HAL1
toggle and also created shortcuts in the app drawer for s2 users
to be able to use quickly instead of having to always go into
Settings > Advanced > LeEco Settings.
------
Change-Id: I2a307a035d5f8cf7e91220b86860bfb004e86d12